+++ packages/tools/atlas/bin/12_world_110m_iso_a2_geojson.js
@@ -6,9 +6,9 @@ import { feature } from 'topojson-client';
 import { writeFile } from '@svizzle/file';
 import { setGeometryPrecision } from '@svizzle/geo';
 
-import topojson from '../data/1/world_110m_iso_a2_topo.json';
+import topojson from '../data/1/World_110m_iso_a2_topo.json';
 
-const PATH = path.resolve(__dirname, '../data/2/world_110m_iso_a2_geo.json');
+const PATH = path.resolve(__dirname, '../data/2/World_110m_iso_a2_geo.json');
 const truncateGeojson = setGeometryPrecision(4);
 const geojson = truncateGeojson(feature(topojson, topojson.objects.countries));