+++ packages/tools/atlas/bin/dist_setup.sh
@@ -1,0 +1,5 @@
+#! /usr/bin/env bash
+
+rm -rf gist && \
+mkdir -p gist && \
+git clone git@gist.github.com:b9538f1b7308d1a2f2d54c927116e825.git gist