+++ CHANGELOG.md
@@ -10,9 +10,10 @@
 
 - added components
 
-## `@svizzle/atlas` v0.1.0
+## `@svizzle/atlas` v0.1.0 (next)
 
-- moved geo scripts from /Choropleth
+- moved world boundaries scripts from /Choropleth
+- add NUTS scripts
 
 ## `@svizzle/geo` v0.4.0