+++ packages/tools/atlas/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## `@svizzle/atlas` v0.1.0 (next)
 
-- moved geo scripts from /Choropleth
+- moved world boundaries scripts from /Choropleth
+- add NUTS scripts